pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jiuwenclaw"
version = "0.1.7.post1"
description = "JiuwenClaw"
requires-python = ">=3.11,<3.14"
license = { text = "Apache-2.0" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "chromadb>=1.5.0",
    "ruamel.yaml>=0.18.0",
    "lark-oapi>=1.5.3",
    "loguru>=0.7.3",
    "openjiuwen==0.1.7",
    "pip==26.0.1",
    "pgvector>=0.4.2",
    "websockets>=12.0",
    "psutil>=7.2.2",
    "aiosqlite>=0.22.1",
    "croniter>=2.0.0",
    "python-telegram-bot>=20.0",
    "discord.py>=2.4.0",
    "dingtalk-stream>=0.24.3",
    "wecom-aibot-sdk>=1.0.0",
    "python-socks>=2.8.1",
    "greenlet>=3.3.2",
    "skillnet-ai==0.0.16",
    "mutagen>=1.47.0",
    "google-genai>=1.68.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
    "opentelemetry-exporter-otlp-proto-http>=1.25.0",
]

[project.optional-dependencies]
desktop = ["pywebview>=5.4"]
dev = ["pyinstaller>=6.0.0", "pywebview>=5.4"]

[project.scripts]
jiuwenclaw-app = "jiuwenclaw.app:main"
jiuwenclaw-agentserver = "jiuwenclaw.app_agentserver:main"
jiuwenclaw-gateway = "jiuwenclaw.app_gateway:main"
jiuwenclaw-web = "jiuwenclaw.app_web:main"
jiuwenclaw-start = "jiuwenclaw.start_services:main"
jiuwenclaw-init = "jiuwenclaw.init_workspace:main"
jiuwenclaw-desktop = "jiuwenclaw.desktop_app:main"

[tool.setuptools]
# 入口脚本现在是 jiuwenclaw 包内的一部分，不需要 py-modules
py-modules = []
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["jiuwenclaw*"]

[tool.setuptools.package-data]
"*" = ["LICENSE"]
jiuwenclaw = [
    "web/dist/**/*",
    "web/dist/*",
    # config / env 模板放在 resources 目录
    "resources/**/*",
    "resources/*",
    "resources/.env.template",
]

[tool.setuptools.exclude-package-data]
jiuwenclaw = ["web/node_modules/**", "web/node_modules/*"]
